avoid N getCommit calls when N commits are out of sync with upstream
Each getCommit call ends up as a `git show` subprocess:

https://github.com/microsoft/vscode/blob/65a3d201e6a97b56ad0980bf2078cf6381ce3e52/extensions/git/src/git.ts#L2468

In my Linux kernel repo my `master` branch goes unused for months at a
time so when I check it out it's normally tens of thousands of commits
behind its upstream. Spinning up that many `git show` commands kills the
extension! Not sure of the exact mechanics of why it gets killed, but
even if it didn't get killed, it would be very slow and power-hungry.

Luckily we actually already have most of the necessary information to
show this part of the status view without getting the Commit objects; we
just need to change the `rev-list` command to a `log` command so we can
add the commit subject.

So I've created a new type for just the minimal commit info and updated
a bunch of places to use this. Now the status view renders in under a
second on my home PC, even with >18k commits to render.