Fix range operator used in getCommitRange
The three-dot notation is a symmetric operation returning the difference
between two commits, i.e. all that are reachable from one but not the
other.

This didn't turn out to matter for the usage directly in
`internalMagitStatus` - we only call `getCommitRange` at most once there
anyway because we first check the `.ahead`/`.behind` counters provided
by the VSCode Git API.

However in `pushRemoteStatus` we call it unconditionall for both
"directions", which means the same set of commits show up in the
returned `.head` and `.behind` fields. This leads to the bug described
in https://github.com/kahole/edamagit/pull/257#issuecomment-1559101338.

The fix is just to use the asymmetric range operator: `..`.