Fix hash list parsing in getCommitRange
Looks like when I wrote this code I didn't realise that
`String.prototype.split` never returns an empty array for a non-empty
separator. That is, for an empty input it returns an array with a single
empty string in it.

This means when there is nothing in the specified range, an empty string
is passed to `getCommit`. The upshot is that now that the range
specification bug fixed by the parent commit is fixed,
`pushRemoteStatus` always encounters an error and returns an empty
result, so the `Push: ` view in the UI disappears.

So we just need to explicitly account for the empty-string case here.

Also remove the comment, which seems to have leaked in during
developent, it doesn't apply to the code here (there is a `.slice` call
nearby but that's just to enforce `maxReults`).