Fix ahead/behind polarity in pushRemoteStatus
"Ahead" means "local branch is ahead". This code was written as if it
meant "remote branch is ahead".

Note: in the next commits I'll be dropping this code anyway, but since I
noticed this bug it seems worth capturing it in the history in case
those later commits get reverted or something.